requirements.txt
python-dotenv>=1.0.0
wikipedia>=1.4.0
openai>=1.0.0
